Add modules in functionality.js
